requirements.txt
alembic==1.9.1
async-timeout==4.0.2
bidict==0.22.0
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Migrate==4.0.0
Flask-SocketIO==5.3.2
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
psycopg2==2.9.5
PyJWT==2.6.0
python-dotenv==0.21.0
python-engineio==4.3.4
python-socketio==5.7.2
redis==4.4.0
six==1.16.0
SQLAlchemy==1.4.45
Werkzeug==2.2.2
